+++ user_guide_src/source/changelog.rst
@@ -408,7 +408,7 @@ Release Date: Not Released
 
       -  Changed ``site_url()`` method  to accept an array as well.
       -  Removed internal method ``_assign_to_config()`` and moved its implementation to *CodeIgniter.php* instead.
-	 -  ``item()`` now returns NULL instead of FALSE when the required config item doesn't exist.
+      -  ``item()`` now returns NULL instead of FALSE when the required config item doesn't exist.
 
    -  :doc:`Security Library <libraries/security>` changes include: